+++ marginaleffects/uncertainty.py
@@ -37,10 +37,12 @@ def get_se(J, V):
     return se
 
 
-def get_z_p_ci(df, model, conf_level):
+def get_z_p_ci(df, model, conf_level, hypothesis_null=0):
     if "std_error" not in df.columns:
         return df
-    df = df.with_columns((pl.col("estimate") / pl.col("std_error")).alias("statistic"))
+    df = df.with_columns(
+        ((pl.col("estimate") - float(hypothesis_null)) / pl.col("std_error")).alias("statistic")
+    )
     if hasattr(model, "df_resid") and isinstance(model.df_resid, float):
         dof = model.df_resid
     else:
@@ -53,6 +55,7 @@ def get_z_p_ci(df, model, conf_level):
     df = df.with_columns(
         (pl.col("estimate") + critical_value * pl.col("std_error")).alias("conf_high")
     )
+
     df = df.with_columns(
         pl.col("statistic")
         .apply(lambda x: (2 * (1 - stats.t.cdf(np.abs(x), dof))))